+++ geo/src/algorithm/intersects/line_string.rs
@@ -19,6 +19,7 @@ where
 symmetric_intersects_impl!(Coord<T>, LineString<T>);
 symmetric_intersects_impl!(Line<T>, LineString<T>);
 symmetric_intersects_impl!(Rect<T>, LineString<T>);
+symmetric_intersects_impl!(Triangle<T>, LineString<T>);
 
 // Blanket implementation from LineString<T>
 impl<T, G> Intersects<G> for MultiLineString<T>
@@ -38,3 +39,4 @@ where
 symmetric_intersects_impl!(Point<T>, MultiLineString<T>);
 symmetric_intersects_impl!(Line<T>, MultiLineString<T>);
 symmetric_intersects_impl!(Rect<T>, MultiLineString<T>);
+symmetric_intersects_impl!(Triangle<T>, MultiLineString<T>);