+++ geo/src/algorithm/intersects/polygon.rs
@@ -1,8 +1,9 @@
 use super::{has_disjoint_bboxes, Intersects};
-use crate::utils::{coord_pos_relative_to_ring, CoordPos};
-use crate::BoundingRect;
+use crate::coordinate_position::CoordPos;
+use crate::{BoundingRect, CoordinatePosition};
 use crate::{
     Coord, CoordNum, GeoNum, Line, LineString, MultiLineString, MultiPolygon, Point, Polygon, Rect,
+    Triangle,
 };
 
 impl<T> Intersects<Coord<T>> for Polygon<T>
@@ -10,11 +11,7 @@ where
     T: GeoNum,
 {
     fn intersects(&self, p: &Coord<T>) -> bool {
-        coord_pos_relative_to_ring(*p, self.exterior()) != CoordPos::Outside
-            && self
-                .interiors()
-                .iter()
-                .all(|int| coord_pos_relative_to_ring(*p, int) != CoordPos::Inside)
+        self.coordinate_position(p) != CoordPos::Outside
     }
 }
 symmetric_intersects_impl!(Coord<T>, Polygon<T>);
@@ -44,6 +41,16 @@ where
     }
 }
 symmetric_intersects_impl!(Rect<T>, Polygon<T>);
+
+impl<T> Intersects<Triangle<T>> for Polygon<T>
+where
+    T: GeoNum,
+{
+    fn intersects(&self, rect: &Triangle<T>) -> bool {
+        self.intersects(&rect.to_polygon())
+    }
+}
+symmetric_intersects_impl!(Triangle<T>, Polygon<T>);
 
 impl<T> Intersects<Polygon<T>> for Polygon<T>
 where
@@ -81,6 +88,7 @@ where
 symmetric_intersects_impl!(Point<T>, MultiPolygon<T>);
 symmetric_intersects_impl!(Line<T>, MultiPolygon<T>);
 symmetric_intersects_impl!(Rect<T>, MultiPolygon<T>);
+symmetric_intersects_impl!(Triangle<T>, MultiPolygon<T>);
 symmetric_intersects_impl!(Polygon<T>, MultiPolygon<T>);
 
 #[cfg(test)]